+++ softlight/tests/sl_fullscreen_quad.cpp
@@ -481,7 +481,7 @@ utils::Pointer<SL_SceneGraph> mesh_test_create_context()
     retCode = meshLoader.load("testdata/african_head/african_head.obj");
     assert(retCode != 0);
 
-    retCode = pGraph->import(meshLoader.data());
+    retCode = (int)pGraph->import(meshLoader.data());
     assert(retCode == 0);
 
     // Always make sure the scene graph is updated before rendering
